requirements.txt
python>=3.13.9
streamlit>=1.52.1
requests>=2.32.5
networkx>=3.6.1
pyvis>=0.3.2
google-generativeai>=0.8.6
graphviz>=0.21